backend/requirements.txt
Flask==3.0.0
redis>=4.6.0
rq>=1.16.0,<2.0
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
Flask-Bcrypt==1.0.1
Flask-JWT-Extended==4.6.0
psycopg2-binary==2.9.9
python-dotenv==1.0.0
email-validator==2.1.0
pefile==2023.2.7
lief==0.13.2
python-magic==0.4.27
capstone==5.0.0
pycryptodome==3.20.0
pefile==2023.2.7
lief==0.13.2
Werkzeug==2.3.7




